New 'staged' approach
Run a minimal bootstrap powershell to start interacting with the device.
Then read additional powershell commands from the device, to upgrade the
bootstrap.

stage.sh can be used to prepare and send the second stage. Basically, the
protocol is to send a 2 byte length, followed by that many bytes of Power-
shell, which will be executed.

Predefined variables may be used, most importantly  (the device filestream)
and  (the HID endpoint block size).