Substantial refactoring
Moved more logic into common files, use a ring buffer for keystrokes and piped data.

This should hopefully start to reduce latency, in theory allowing better performance.

Note that the piped data is only sent to the PC after the first write is received,
indicating that the PC is ready to receive it.