Working multithreaded Powershell second stage
Needed to pass parameters for  and  (a.k.a) , since the thread uses
a new scope, it seems.

Also add a bit more robustness into read_exec, in case it does not get the data required
in the first loop through. This could still fail if only 1 length byte makes it through
on one read. Unlikely, I guess, but possible.

Also begin to implement flow control in the msf stage, to avoid flooding the AVR