Implement flow control in the Generic HID "pipe"
In the process, enable interrupt driven UART Transmission.

In order to check whether the victim is permitted to write to the
pipe, the victim should first do a write to the device. It is
permissble to write an empty packet, i.e. one with the first byte
of the packet == 0.

Having written the packet, the subsequent read will have the length
byte OR'd with (fc << 7). i.e. bit 7 will indicate whether the
pipe is blocked or not. If set, the pipe is blocked, and the victim
should not write. Otherwise, the victim may write a single packet.

NB: In order to read a packet from the device, it is imperative to
first write a packet. This is the only way that the device can know
whether the victim is ready to read or not!