Support specifying the device to flash.
In case there are multiple USB serial interfaces that match the device pattern,
it is now possible to override the default DEVICEPATTERN on the command line, or
with an environment variable.

For example:

DEVICEPATTERN=/dev/cu.usbmodem14241 ./flash_avr ...

Also make the messages a little more verbose, so that using the incorrect device
can be noticed.