Add a linux-based VNC server
This emits the keystrokes used to type the first stage.

NB:

While it is quite possible to type out the payload one keystroke
at a time, it is significantly more efficient to use the 'pastefile'
command to transfer the entire payload in one go, and eliminate
all of the latency.

NB2:

This VNC server does not yet implement the mouse jiggler, but it should
be a relatively simple thing to implement (famous last words!)